Sqbot model files
