+++ squidle/yolov5m/label_map.json
@@ -1,0 +1,4 @@
+[
+    null,
+    [{"or":[{"name":"vocab_elements","op":"any","val":{"name":"key","op":"eq","val":"28382"}},{"name":"vocab_elements","op":"any","val":{"name":"key","op":"eq","val":"13029"}}]}]
+]
